+++ CHANGELOG.md
@@ -5,6 +5,9 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Plotting function `residual_plot` to create plots of residuals vs regression
+parameters. 
 
 [0.13.4]: https://github.com/pvcaptest/pvcaptest/compare/v0.13.3...v0.13.4
 ## [0.13.4] - 2026-02-17